+++ tests/PHPUnit/Integration/Tracker/ActionTest.php
@@ -13,6 +13,7 @@ use Piwik\Plugins\SitesManager\API;
 use Piwik\Tests\Framework\Fixture;
 use Piwik\Tests\Framework\Mock\FakeAccess;
 use Piwik\Tracker\Action;
+use Piwik\Tracker\ActionPageview;
 use Piwik\Tracker\PageUrl;
 use Piwik\Tracker\Request;
 use Piwik\Plugin\Manager as PluginManager;
@@ -35,7 +36,7 @@ class ActionTest extends IntegrationTestCase
         $section['campaign_keyword_var_name']     = 'piwik_kwd,matomo_kwd,utm_term,test_piwik_kwd';
         Config::getInstance()->Tracker = $section;
 
-        PluginManager::getInstance()->loadPlugins(array('SitesManager'));
+        PluginManager::getInstance()->loadPlugins(array('Actions', 'SitesManager'));
 
         Fixture::loadAllTranslations();
     }
@@ -50,6 +51,39 @@ class ActionTest extends IntegrationTestCase
     protected function setUpRootAccess()
     {
         FakeAccess::$superUser = true;
+    }
+
+    public function test_isCustomActionRequest()
+    {
+        $request = new Request(array('ca' => '1'));
+        $this->assertTrue(Action::isCustomActionRequest($request));
+
+        $request = new Request(array('ca' => '0'));
+        $this->assertFalse(Action::isCustomActionRequest($request));
+
+        $request = new Request(array());
+        $this->assertFalse(Action::isCustomActionRequest($request));
+    }
+
+    public function test_factory_notDefaultsToPageViewWhenCustomPluginRequest()
+    {
+        $this->expectException(\Exception::class);
+        $this->expectExceptionMessage('Request was meant for a plugin which is no longer activated. Request needs to be ignored.');
+        $this->setUpRootAccess();
+        $idSite = API::getInstance()->addSite("site1", array('http://example.org'));
+        $request = new Request(array('ca' => '1', 'idsite' => $idSite));
+
+        Action::factory($request);
+    }
+
+    public function test_factory_defaultsToPageviewWhenNotCustomPluginRequest()
+    {
+        $this->setUpRootAccess();
+        $idSite = API::getInstance()->addSite("site1", array('http://example.org'));
+        $request = new Request(array('idsite' => $idSite));
+
+        $action =  Action::factory($request);
+        $this->assertTrue($action instanceof ActionPageview);
     }
 
     public function getTestUrls()
@@ -383,7 +417,6 @@ class ActionTest extends IntegrationTestCase
      */
     public function testExtractUrlAndActionNameFromRequest($request, $expected)
     {
-        PluginManager::getInstance()->loadPlugins(array('Actions', 'SitesManager'));
         $this->setUpRootAccess();
         $idSite = API::getInstance()->addSite("site1", array('http://example.org'));
         $request['idsite'] = $idSite;
@@ -397,7 +430,7 @@ class ActionTest extends IntegrationTestCase
           'type' => $action->getActionType(),
         );
 
-        $this->assertEquals($processed, $expected);
+        $this->assertEquals($expected, $processed);
     }
 
     public function provideContainerConfig()